+++ src/components/NewsletterSignup/newsletterSignupChangelogData.ts
@@ -17,7 +17,9 @@ export const changelogData: ChangelogData[] = [
     notes: [
       "Removes explicit `className` prop as interface can be extended to include Chakra prop or HTML attribute types.",
       "Replaces `highlightColorTypesArray` with shared highlight colors array plus default gray, now `newsletterHighlightColorsArray`. ",
-      "Updated the internal use of ids to reduce `undefined` ids in the DOM.",
+      "Updates the internal use of ids to reduce `undefined` ids in the DOM.",
+      "Removes the internal `newsletter-signup` classes and replaces them with data-testids.",
+      "Adds `data-testid` value of `ds-newsletterSignup` to the parent element.",
     ],
   },
   {