+++ src/components/SocialMediaLinks/socialMediaLinksChangelogData.ts
@@ -17,7 +17,9 @@ export const changelogData: ChangelogData[] = [
     notes: [
       "Removes explicit `className` prop as interface can be extended to include Chakra prop or HTML attribute types.",
       "Replaces `sizeTypeArray` with shared sizes array.",
-      "Updated the internal use of ids to reduce `undefined` ids in the DOM.",
+      "Updates the internal use of ids to reduce `undefined` ids in the DOM.",
+      "Renames the className `platLink` to `ds-socialMediaLinks-icon` to align with the new naming convention.",
+      "Adds `data-testid` value of `ds-socialMediaLinks` to the parent element.",
     ],
   },
   {